Move all progress/step-related code into a 'progress' dir.
