+++ frontend/lib/hpaction/emergency-hp-action.tsx
@@ -1,13 +1,13 @@
 import React, { useContext } from 'react';
 import Routes, { getSignupIntentOnboardingInfo } from "../routes";
-import { ProgressRoutesProps, buildProgressRoutesComponent } from '../progress-routes';
+import { ProgressRoutesProps, buildProgressRoutesComponent } from '../progress/progress-routes';
 import { HPUploadStatus, OnboardingInfoSignupIntent, HPDocusignStatus } from "../queries/globalTypes";
 import Page from '../page';
 import { GetStartedButton } from '../get-started-button';
 import { AppContext } from '../app-context';
 import { TenantChildren } from './hp-action-tenant-children';
 import { isNotSuingForRepairs } from './hp-action-util';
-import { MiddleProgressStep, ProgressStepProps } from '../progress-step-route';
+import { MiddleProgressStep, ProgressStepProps } from '../progress/progress-step-route';
 import { ProgressButtons, BackButton, NextButton } from '../buttons';
 import { Link } from 'react-router-dom';
 import { EhpAccessForInspection } from './hp-action-access-for-inspection';
@@ -32,7 +32,7 @@ import { Formset } from '../forms/formset';
 import { CUSTOM_ISSUE_MAX_LENGTH, MAX_CUSTOM_ISSUES_PER_AREA } from '../../../common-data/issue-validation.json';
 import { FormsetItem, formsetItemProps } from '../forms/formset-item';
 import { TextualFieldWithCharsRemaining } from '../forms/chars-remaining';
-import { SessionStepBuilder } from '../session-step-builder';
+import { SessionStepBuilder } from '../progress/session-step-builder';
 import { OptionalLandlordDetailsMutation } from '../queries/OptionalLandlordDetailsMutation';
 import { PhoneNumberFormField } from '../forms/phone-number-form-field';
 import { isUserNycha } from '../nycha';
