+++ frontend/lib/hpaction/fee-waiver.tsx
@@ -10,12 +10,12 @@ import { FeeWaiverMiscMutation } from '../queries/FeeWaiverMiscMutation';
 import { FeeWaiverIncomeMutation } from '../queries/FeeWaiverIncomeMutation';
 import { FeeWaiverExpensesMutation, BlankFeeWaiverExpensesInput } from '../queries/FeeWaiverExpensesMutation';
 import { ProgressiveOtherCheckboxFormField } from '../forms/other-checkbox-form-field';
-import { MiddleProgressStep } from '../progress-step-route';
+import { MiddleProgressStep } from '../progress/progress-step-route';
 import { FeeWaiverPublicAssistanceMutation } from '../queries/FeeWaiverPublicAssistanceMutation';
 import { Link } from 'react-router-dom';
 import Routes from '../routes';
 import { bulmaClasses } from '../bulma';
-import { SessionStepBuilder } from '../session-step-builder';
+import { SessionStepBuilder } from '../progress/session-step-builder';
 
 const INITIAL_EXPENSES_STATE: FeeWaiverExpensesInput = {
   ...BlankFeeWaiverExpensesInput,
