+++ frontend/lib/hpaction/hp-action.tsx
@@ -6,17 +6,17 @@ import { ProgressButtons } from '../buttons';
 import { IssuesRoutes } from '../issues/issue-pages';
 import { withAppContext, AppContextType } from '../app-context';
 import { PdfLink } from '../pdf-link';
-import { ProgressRoutesProps, buildProgressRoutesComponent } from '../progress-routes';
+import { ProgressRoutesProps, buildProgressRoutesComponent } from '../progress/progress-routes';
 import { OutboundLink } from '../analytics/google-analytics';
 import { HPUploadStatus, OnboardingInfoSignupIntent } from '../queries/globalTypes';
 import { FeeWaiverMisc, FeeWaiverIncome, FeeWaiverExpenses, FeeWaiverPublicAssistance, FeeWaiverStart } from './fee-waiver';
-import { MiddleProgressStep } from '../progress-step-route';
+import { MiddleProgressStep } from '../progress/progress-step-route';
 import { TenantChildren } from './hp-action-tenant-children';
 import { createHPActionPreviousAttempts } from './hp-action-previous-attempts';
 import { CheckboxFormField } from '../forms/form-fields';
 import { HpActionUrgentAndDangerousMutation } from '../queries/HpActionUrgentAndDangerousMutation';
 import { YesNoRadiosFormField } from '../forms/yes-no-radios-form-field';
-import { SessionStepBuilder } from '../session-step-builder';
+import { SessionStepBuilder } from '../progress/session-step-builder';
 import { HarassmentApartment, HarassmentExplain, HarassmentAllegations1, HarassmentAllegations2 } from './hp-action-harassment';
 import { HpActionSueMutation } from '../queries/HpActionSueMutation';
 import { HarassmentCaseHistory } from './hp-action-case-history';
