+++ frontend/lib/loc/access-dates.tsx
@@ -9,7 +9,7 @@ import { ProgressButtons } from "../buttons";
 import { dateAsISO, addDays } from '../util';
 
 import validation from '../../../common-data/access-dates-validation.json';
-import { MiddleProgressStep } from '../progress-step-route';
+import { MiddleProgressStep } from '../progress/progress-step-route';
 
 /**
  * The minimum number of days from today that the first access date
