+++ frontend/lib/onboarding/onboarding-routes.tsx
@@ -4,10 +4,10 @@ import { Route, Switch } from 'react-router';
 import OnboardingStep1 from './onboarding-step-1';
 import OnboardingStep3 from './onboarding-step-3';
 import OnboardingStep4 from './onboarding-step-4';
-import { RouteProgressBar } from '../progress-bar';
-import { RedirectToLatestStep } from '../progress-redirection';
+import { RouteProgressBar } from '../progress/progress-bar';
+import { RedirectToLatestStep } from '../progress/progress-redirection';
 import { OnboardingInfoSignupIntent } from '../queries/globalTypes';
-import { ProgressStepRoute } from '../progress-step-route';
+import { ProgressStepRoute } from '../progress/progress-step-route';
 import { OnboardingThanks } from './onboarding-thanks';
 
 export type OnboardingRoutesProps = {
