+++ frontend/lib/onboarding/relief-attempts.tsx
@@ -6,7 +6,7 @@ import { ReliefAttemptsMutation } from '../queries/ReliefAttemptsMutation';
 import { YesNoRadiosFormField } from '../forms/yes-no-radios-form-field';
 import { ProgressButtons } from '../buttons';
 import { toStringifiedNullBool } from '../forms/form-input-converter';
-import { MiddleProgressStep } from '../progress-step-route';
+import { MiddleProgressStep } from '../progress/progress-step-route';
 
 
 const ReliefAttemptsPage = MiddleProgressStep(props => (
