+++ frontend/lib/progress/progress-step-route.tsx
@@ -2,9 +2,9 @@ import React, { useContext } from 'react';
 
 import { RouteComponentProps, Route } from "react-router";
 import { getRelativeStep } from "./progress-util";
-import { AllSessionInfo } from './queries/AllSessionInfo';
-import { AppContext } from './app-context';
-import { assertNotNull } from './util';
+import { AllSessionInfo } from '../queries/AllSessionInfo';
+import { AppContext } from '../app-context';
+import { assertNotNull } from '../util';
 
 export type BaseProgressStepRoute = {
   /** The route's URL path. */
