+++ frontend/lib/progress/progress-util.ts
