+++ frontend/lib/rh/rental-history.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useRef } from 'react';
 
-import { ProgressRoutesProps, buildProgressRoutesComponent } from "../progress-routes";
+import { ProgressRoutesProps, buildProgressRoutesComponent } from "../progress/progress-routes";
 import Routes from "../routes";
 import Page from "../page";
 import { StaticImage } from '../static-image';
